Squashed commit of the following:
commit 0ad78523ef9770de64ea7ddb7bc8a36b17909208
Author: sbinalla <92679973+youreondrugs@users.noreply.github.com>
Date:   Tue Mar 19 18:40:55 2024 +0200

    Pvp 102 (#41)


commit b1b0726931ad4ffe711542bf161dbe3e66b83099
Author: Vytautas Jaras <vjaras202@gmail.com>
Date:   Mon Mar 18 20:35:24 2024 +0200

    PVP-166 (#47)
