+++ app/src/main/java/com/pvp/app/ui/screen/calendar/CalendarMonthlyViewModel.kt
@@ -20,12 +20,11 @@ import java.time.YearMonth
 import javax.inject.Inject
 
 @HiltViewModel
-class MonthlyCalendarViewModel @Inject constructor(
+class CalendarMonthlyViewModel @Inject constructor(
     private val taskService: TaskService,
     private val userService: UserService
 ) : ViewModel() {
 
-    private val _currentYearMonth = MutableStateFlow(YearMonth.now())
     private val _state = MutableStateFlow(CalendarUiState.Init)
     val state: StateFlow<CalendarUiState> = _state.asStateFlow()
