+++ app/src/main/java/com/pvp/app/ui/screen/calendar/CalendarWeeklyViewModel.kt
@@ -29,10 +29,10 @@ import javax.inject.Inject
 
 @HiltViewModel
 @OptIn(ExperimentalCoroutinesApi::class)
-class CalendarViewModel @Inject constructor(
+class CalendarWeeklyViewModel @Inject constructor(
     private val taskService: TaskService,
     private val userService: UserService,
-    private val healtConnectService: HealthConnectService,
+    private val healthConnectService: HealthConnectService,
     private val client: HealthConnectClient
 ) : ViewModel() {
 
@@ -85,7 +85,7 @@ class CalendarViewModel @Inject constructor(
             .atStartOfDay(ZoneId.systemDefault())
             .toInstant()
 
-        return healtConnectService.aggregateSteps(
+        return healthConnectService.aggregateSteps(
             start,
             end
         )
