+++ app/src/main/java/com/pvp/app/ui/screen/steps/StepScreen.kt
@@ -4,12 +4,13 @@ import android.os.Build
 import androidx.activity.compose.rememberLauncherForActivityResult
 import androidx.annotation.RequiresApi
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -23,10 +24,68 @@ import androidx.health.connect.client.PermissionController
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 
+@Composable
+fun ActivityList(
+    model: StepViewModel = hiltViewModel()
+) {
+    // Required for checking whether user has permissions before entering the window,
+    // as users can revoke permissions at any time
+    val permissionContract = PermissionController.createRequestPermissionResultContract()
+
+    val launcher =
+        rememberLauncherForActivityResult(permissionContract) {
+            model.getExercises()
+        }
+
+    LaunchedEffect(Unit) {
+        if (model.permissionsGranted()) {
+            model.getExercises()
+        } else {
+            launcher.launch(PERMISSIONS)
+        }
+    }
+
+    val activities = model.activites
+        .collectAsStateWithLifecycle().value
+        .groupingBy { it.title }
+        .eachCount()
+        .toList()
+        .sortedByDescending { it.second }
+
+
+    Column(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp)
+    ) {
+        Text(
+            text = "Your favorite activities:",
+            textAlign = TextAlign.Center,
+            fontSize = 24.sp,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(bottom = 16.dp)
+        )
+
+        Spacer(modifier = Modifier.height((16.dp)))
+
+        activities.forEachIndexed { index, activity ->
+            Text(
+                text = "${index + 1}. ${activity.first} (${activity.second} occurences)",
+                textAlign = TextAlign.Start,
+                fontSize = 18.sp,
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .padding(bottom = 8.dp)
+            )
+        }
+    }
+}
+
 @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
 @Composable
 fun StepCounter(
-    viewModel: StepViewModel = hiltViewModel()
+    model: StepViewModel = hiltViewModel()
 ) {
 
     // Required for checking whether user has permissions before entering the window,
@@ -35,18 +94,18 @@ fun StepCounter(
 
     val launcher =
         rememberLauncherForActivityResult(permissionContract) {
-            viewModel.updateTodaysSteps()
+            model.updateTodaysSteps()
         }
 
     LaunchedEffect(Unit) {
-        if (viewModel.permissionsGranted()) {
-            viewModel.updateTodaysSteps()
+        if (model.permissionsGranted()) {
+            model.updateTodaysSteps()
         } else {
-            launcher.launch(viewModel.PERMISSIONS)
+            launcher.launch(PERMISSIONS)
         }
     }
 
-    val steps = viewModel.stepsCount.collectAsStateWithLifecycle()
+    val steps = model.stepsCount.collectAsStateWithLifecycle()
 
     Text(
         steps.value.toString(),
@@ -67,7 +126,6 @@ fun StepScreen() {
             .fillMaxSize()
             .background(MaterialTheme.colorScheme.surface)
             .padding(8.dp),
-        verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
         Text(
@@ -79,5 +137,13 @@ fun StepScreen() {
         Spacer(modifier = Modifier.height((16.dp)))
 
         StepCounter()
+
+        Spacer(modifier = Modifier.height((16.dp)))
+
+        HorizontalDivider()
+
+        Spacer(modifier = Modifier.height((16.dp)))
+
+        ActivityList()
     }
 }
