+++ app/src/main/java/com/pvp/app/ui/screen/steps/StepViewModel.kt
@@ -9,6 +9,7 @@ import androidx.health.connect.client.records.StepsRecord
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.pvp.app.api.HealthConnectService
+import com.pvp.app.model.SportActivity
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
@@ -24,13 +25,11 @@ class StepViewModel @Inject constructor(
     private val service: HealthConnectService
 ) : ViewModel() {
 
-    val PERMISSIONS = setOf(
-        HealthPermission.getReadPermission(StepsRecord::class),
-        HealthPermission.getReadPermission(ExerciseSessionRecord::class)
-    )
-
-    private val _stepsCount = MutableStateFlow(0)
+    private val _stepsCount = MutableStateFlow(0L)
     val stepsCount = _stepsCount.asStateFlow()
+
+    private val _activities = MutableStateFlow<List<SportActivity>>(emptyList())
+    val activites = _activities.asStateFlow()
 
     suspend fun permissionsGranted(): Boolean {
         val granted = client.permissionController.getGrantedPermissions()
@@ -41,21 +40,48 @@ class StepViewModel @Inject constructor(
     @RequiresApi(Build.VERSION_CODES.UPSIDE_DOWN_CAKE)
     fun updateTodaysSteps() {
         viewModelScope.launch {
+            val end = LocalDate
+                .now()
+                .plusDays(1)
+                .atStartOfDay(ZoneId.systemDefault())
+                .toInstant()
+
+            val start = LocalDate
+                .now()
+                .atStartOfDay(ZoneId.systemDefault())
+                .toInstant()
+
+            _stepsCount.value = service.aggregateSteps(
+                start = start,
+                end = end
+            )
+        }
+    }
+
+    fun getExercises() {
+        viewModelScope.launch {
             val end = Instant.now()
-            val start = LocalDate.now().atStartOfDay(ZoneId.systemDefault()).toInstant()
 
-            try {
-                val records: List<StepsRecord> = service.readActivityData(
-                    record = StepsRecord::class,
-                    start = start,
-                    end = end
-                )
+            val start = LocalDate
+                .now()
+                .minusDays(29)
+                .atStartOfDay(ZoneId.systemDefault())
+                .toInstant()
 
-                _stepsCount.value = records.sumOf { it.count.toInt() }
-            } catch (e: Exception) {
-                // Exception should be handled by showing a Toast
-                e.printStackTrace()
+            val records = service.readActivityData(
+                record = ExerciseSessionRecord::class,
+                start = start,
+                end = end
+            )
+
+            _activities.value = records.mapNotNull { record ->
+                SportActivity.fromId(record.exerciseType)
             }
         }
     }
 }
+
+val PERMISSIONS = setOf(
+    HealthPermission.getReadPermission(StepsRecord::class),
+    HealthPermission.getReadPermission(ExerciseSessionRecord::class)
+)
