+++ app/src/main/res/values/strings.xml
@@ -16,7 +16,8 @@
     <string name="route_authentication_sign_up">Sign Up</string>
     <string name="route_survey">Survey</string>
     <string name="route_calendar">Calendar</string>
-    <string name="route_calendar_monthly">Monthly calendar</string>
+    <string name="route_filters_edit_activities">Edit activities filter</string>
+    <string name="route_filters_edit_ingredients">Edit ingredients filter</string>
     <string name="route_profile">Profile</string>
     <string name="route_settings">Settings</string>
     <string name="route_steps">Steps</string>
